requirements.txt
# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0

# OpenAI Integration
openai>=1.0.0

# Web Framework Dependencies
python-multipart==0.0.6
pydantic==2.5.0
python-dotenv==1.0.0

# Database
# Using SQLite (built-in with Python)

# Security & Rate Limiting
slowapi==0.1.9

# PDF Generation
reportlab==4.0.7

# Development & Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2

# Optional: For production deployment
gunicorn==21.2.0